refactor: split the tests into multiple files
