+++ __tests__/main.spec.ts
@@ -1,70 +1,9 @@
 import * as github from '@actions/github';
 import {Issue} from '../src/classes/issue';
-
 import {IssuesProcessor} from '../src/classes/issues-processor';
 import {IIssuesProcessorOptions} from '../src/interfaces/issues-processor-options';
-import {IsoDateString} from '../src/types/iso-date-string';
-
-function generateIssue(
-  options: IIssuesProcessorOptions,
-  id: number,
-  title: string,
-  updatedAt: IsoDateString,
-  createdAt: IsoDateString = updatedAt,
-  isPullRequest = false,
-  labels: string[] = [],
-  isClosed = false,
-  isLocked = false,
-  milestone = ''
-): Issue {
-  return new Issue(options, {
-    number: id,
-    labels: labels.map(l => {
-      return {name: l};
-    }),
-    title,
-    created_at: createdAt,
-    updated_at: updatedAt,
-    pull_request: isPullRequest ? {} : null,
-    state: isClosed ? 'closed' : 'open',
-    locked: isLocked,
-    milestone: {
-      title: milestone
-    }
-  });
-}
-
-const DefaultProcessorOptions: IIssuesProcessorOptions = Object.freeze({
-  repoToken: 'none',
-  staleIssueMessage: 'This issue is stale',
-  stalePrMessage: 'This PR is stale',
-  closeIssueMessage: 'This issue is being closed',
-  closePrMessage: 'This PR is being closed',
-  daysBeforeStale: 1,
-  daysBeforeIssueStale: NaN,
-  daysBeforePrStale: NaN,
-  daysBeforeClose: 30,
-  daysBeforeIssueClose: NaN,
-  daysBeforePrClose: NaN,
-  staleIssueLabel: 'Stale',
-  closeIssueLabel: '',
-  exemptIssueLabels: '',
-  stalePrLabel: 'Stale',
-  closePrLabel: '',
-  exemptPrLabels: '',
-  onlyLabels: '',
-  operationsPerRun: 100,
-  debugOnly: true,
-  removeStaleWhenUpdated: false,
-  ascending: false,
-  skipStaleIssueMessage: false,
-  skipStalePrMessage: false,
-  deleteBranch: false,
-  startDate: '',
-  exemptMilestones: '',
-  exemptIssueMilestones: '',
-  exemptPrMilestones: ''
-});
+import {DefaultProcessorOptions} from './constants/default-processor-options';
+import {generateIssue} from './functions/generate-issue';
 
 test('empty issue list results in 1 operation', async () => {
   const processor = new IssuesProcessor(
@@ -1989,460 +1928,4 @@ test('an issue with an exempted milestone and with an exempted issue milestone w
   expect(processor.staleIssues.length).toStrictEqual(0);
   expect(processor.closedIssues.length).toStrictEqual(0);
   expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR without a milestone will be marked as stale', async () => {
-  expect.assertions(3);
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      DefaultProcessorOptions,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      ''
-    )
-  ];
-  const processor = new IssuesProcessor(
-    DefaultProcessorOptions,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(1);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR without an exempted milestone will be marked as stale', async () => {
-  expect.assertions(3);
-  const opts = {...DefaultProcessorOptions};
-  opts.exemptMilestones = 'Milestone1';
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      'Milestone'
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(1);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR with an exempted milestone will not be marked as stale', async () => {
-  expect.assertions(3);
-  const opts = {...DefaultProcessorOptions};
-  opts.exemptMilestones = 'Milestone1';
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      'Milestone1'
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(0);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR with an exempted milestone will not be marked as stale (multi milestones with spaces)', async () => {
-  expect.assertions(3);
-  const opts = {...DefaultProcessorOptions};
-  opts.exemptMilestones = 'Milestone1, Milestone2';
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      'Milestone2'
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(0);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR with an exempted milestone will not be marked as stale (multi milestones without spaces)', async () => {
-  expect.assertions(3);
-  const opts = {...DefaultProcessorOptions};
-  opts.exemptMilestones = 'Milestone1,Milestone2';
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      'Milestone2'
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(0);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR with an exempted milestone but without an exempted issue milestone will not be marked as stale', async () => {
-  expect.assertions(3);
-  const opts = {...DefaultProcessorOptions};
-  opts.exemptMilestones = 'Milestone1';
-  opts.exemptPrMilestones = '';
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      'Milestone1'
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(0);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR with an exempted milestone but with another exempted issue milestone will be marked as stale', async () => {
-  expect.assertions(3);
-  const opts = {...DefaultProcessorOptions};
-  opts.exemptMilestones = 'Milestone1';
-  opts.exemptPrMilestones = 'Milestone2';
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      'Milestone1'
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(1);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('a PR with an exempted milestone and with an exempted issue milestone will not be marked as stale', async () => {
-  expect.assertions(3);
-  const opts = {...DefaultProcessorOptions};
-  opts.exemptMilestones = 'Milestone1';
-  opts.exemptPrMilestones = 'Milestone1';
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'My first issue',
-      '2020-01-01T17:00:00Z',
-      '2020-01-01T17:00:00Z',
-      true,
-      undefined,
-      undefined,
-      undefined,
-      'Milestone1'
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num: number, dt: string) => [],
-    async (issue: Issue, label: string) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toStrictEqual(0);
-  expect(processor.closedIssues.length).toStrictEqual(0);
-  expect(processor.removedLabelIssues.length).toStrictEqual(0);
-});
-
-test('processing an issue opened since 2 days and with the option "daysBeforeIssueStale" at 3 will not make it stale', async () => {
-  expect.assertions(2);
-  const opts: IIssuesProcessorOptions = {
-    ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 3
-  };
-  const issueDate = new Date();
-  issueDate.setDate(issueDate.getDate() - 2);
-  const TestIssueList: Issue[] = [
-    generateIssue(opts, 1, 'An issue with no label', issueDate.toDateString())
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num, dt) => [],
-    async (issue, label) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toEqual(0);
-  expect(processor.closedIssues.length).toEqual(0);
-});
-
-test('processing an issue opened since 2 days and with the option "daysBeforeIssueStale" at 2 will make it stale', async () => {
-  expect.assertions(2);
-  const opts: IIssuesProcessorOptions = {
-    ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 2
-  };
-  const issueDate = new Date();
-  issueDate.setDate(issueDate.getDate() - 2);
-  const TestIssueList: Issue[] = [
-    generateIssue(opts, 1, 'An issue with no label', issueDate.toDateString())
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num, dt) => [],
-    async (issue, label) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toEqual(1);
-  expect(processor.closedIssues.length).toEqual(0);
-});
-
-test('processing an issue opened since 2 days and with the option "daysBeforeIssueStale" at 1 will make it stale', async () => {
-  expect.assertions(2);
-  const opts: IIssuesProcessorOptions = {
-    ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforeIssueStale: 1
-  };
-  const issueDate = new Date();
-  issueDate.setDate(issueDate.getDate() - 2);
-  const TestIssueList: Issue[] = [
-    generateIssue(opts, 1, 'An issue with no label', issueDate.toDateString())
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num, dt) => [],
-    async (issue, label) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toEqual(1);
-  expect(processor.closedIssues.length).toEqual(0);
-});
-
-test('processing a pull request opened since 2 days and with the option "daysBeforePrStale" at 3 will not make it stale', async () => {
-  expect.assertions(2);
-  const opts: IIssuesProcessorOptions = {
-    ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 3
-  };
-  const issueDate = new Date();
-  issueDate.setDate(issueDate.getDate() - 2);
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'A pull request with no label',
-      issueDate.toDateString(),
-      issueDate.toDateString(),
-      true
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num, dt) => [],
-    async (issue, label) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toEqual(0);
-  expect(processor.closedIssues.length).toEqual(0);
-});
-
-test('processing a pull request opened since 2 days and with the option "daysBeforePrStale" at 2 will make it stale', async () => {
-  expect.assertions(2);
-  const opts: IIssuesProcessorOptions = {
-    ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 2
-  };
-  const issueDate = new Date();
-  issueDate.setDate(issueDate.getDate() - 2);
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'A pull request with no label',
-      issueDate.toDateString(),
-      issueDate.toDateString(),
-      true
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num, dt) => [],
-    async (issue, label) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toEqual(1);
-  expect(processor.closedIssues.length).toEqual(0);
-});
-
-test('processing a pull request opened since 2 days and with the option "daysBeforePrStale" at 1 will make it stale', async () => {
-  expect.assertions(2);
-  const opts: IIssuesProcessorOptions = {
-    ...DefaultProcessorOptions,
-    daysBeforeStale: 10,
-    daysBeforePrStale: 1
-  };
-  const issueDate = new Date();
-  issueDate.setDate(issueDate.getDate() - 2);
-  const TestIssueList: Issue[] = [
-    generateIssue(
-      opts,
-      1,
-      'A pull request with no label',
-      issueDate.toDateString(),
-      issueDate.toDateString(),
-      true
-    )
-  ];
-  const processor = new IssuesProcessor(
-    opts,
-    async () => 'abot',
-    async p => (p == 1 ? TestIssueList : []),
-    async (num, dt) => [],
-    async (issue, label) => new Date().toDateString()
-  );
-
-  // process our fake issue list
-  await processor.processIssues(1);
-
-  expect(processor.staleIssues.length).toEqual(1);
-  expect(processor.closedIssues.length).toEqual(0);
 });
