+++ tsconfig.app.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["node_modules", "**/*.spec.ts"],
+  "include": ["src"]
+}
