+++ tsconfig.spec.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["src", "__tests__"],
+  "exclude": ["node_modules"]
+}
